requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
streamlit>=1.35.0

# ─── LLM Providers ────────────────────────────────────────────────────────────
openai>=1.30.0
groq>=0.9.0
google-generativeai>=0.5.0

# ─── RAG / Embeddings ─────────────────────────────────────────────────────────
sentence-transformers>=2.7.0
numpy>=1.26.0

# ─── Document Processing ──────────────────────────────────────────────────────
pdfplumber>=0.11.0
python-docx>=1.1.0

# ─── Web Search ───────────────────────────────────────────────────────────────
tavily-python>=0.3.0
duckduckgo-search>=6.1.0
requests>=2.31.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
